Move LND specific rpc calls from tasks to LNNode
